+++ src/domain/services/imports/get-user-import-by-id.spec.ts
@@ -1,53 +1,14 @@
 import { describe, expect, it } from 'vitest'
 
-import { makeUser } from '@/test/factories/make-user'
-
 import { makeUserImport } from '@/test/factories/make-user-import'
-import { GetUserImport } from './get-user-import-by-id'
-import { makeManyRawImportSeries } from '@/test/factories/make-import-series'
+import { getUserImportById } from './get-user-import-by-id'
 
-describe('get user import', () => {
+describe('get user import by id', () => {
   it('should be able to get user import by id', async () => {
-    const { id: userId } = await makeUser({})
+    const userImport = await makeUserImport({})
 
-    const series = makeManyRawImportSeries(4, {})
+    const sut = await getUserImportById(userImport.id)
 
-    const result = await makeUserImport({
-      userId,
-      series,
-    })
-
-    const sut = await GetUserImport(result.id)
-
-    expect(sut).toEqual(result)
-  })
-
-  it('should be able to get user import by id when movies are empty', async () => {
-    const { id: userId } = await makeUser({})
-
-    const series = makeManyRawImportSeries(4, {})
-
-    const result = await makeUserImport({
-      userId,
-      series,
-      movies: [],
-    })
-
-    const sut = await GetUserImport(result.id)
-
-    expect(sut).toEqual(result)
-  })
-
-  it('should be able to get user import by id when series are empty', async () => {
-    const { id: userId } = await makeUser({})
-
-    const result = await makeUserImport({
-      userId,
-      series: [],
-    })
-
-    const sut = await GetUserImport(result.id)
-
-    expect(sut).toEqual(result)
+    expect(sut.id).toEqual(userImport.id)
   })
 })